face-recognition/requirements.txt
redis
python-dotenv
face-recognition==1.3.0
flask
numpy==1.21.0